Add modesl with flask-migrate
