+++ .gitignore
@@ -1,8 +1,4 @@
 .cache
-
 .env
-
-# frontend
-app/__pycache__
-app/spotify/__pycache__
 venv
+__pycache__
