+++ app/requirements.txt
@@ -1,0 +1,26 @@
+aiohttp==3.8.3
+aiosignal==1.2.0
+alembic==1.8.1
+async-timeout==4.0.2
+attrs==22.1.0
+certifi==2022.9.24
+charset-normalizer==2.1.1
+click==8.1.3
+colorama==0.4.6
+Flask==2.2.2
+Flask-Migrate==3.1.0
+Flask-SQLAlchemy==3.0.2
+frozenlist==1.3.1
+greenlet==1.1.3.post0
+idna==3.4
+itsdangerous==2.1.2
+Jinja2==3.1.2
+Mako==1.2.3
+MarkupSafe==2.1.1
+multidict==6.0.2
+python-dotenv==0.21.0
+requests==2.28.1
+SQLAlchemy==1.4.42
+urllib3==1.26.12
+Werkzeug==2.2.2
+yarl==1.8.1
